OhMyCards

Documentation for OhMyCards.

OhMyCards logo

Making a logo with Makie.jl

OhMyCards.ExampleProcessing

OhMyCards.ExampleProcessing — Type

ExampleProcessing <: Documenter.Builder.DocumenterPipeline

What does this do?

Moves Cardmeta blocks

Adds a quick example block (in Vitepress syntax) if requested in the pipeline and a cardmeta block is found

Adds badges to the page if necessary

Adds the page to expandfirst if it is not already there

This is run at priority 1.2, meaning after doctest and before expand_templates.